# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian API",
    "version": "0.7.0",
    "description": "Everything the web panel does, as JSON over HTTP. Authenticate with an API token (Authorization: Bearer mrd_...) or the supervisor's browser session. Users who sign in to their own page reach only the User page operations. Errors are {\"error\": \"...\"} with a 4xx/5xx status."
  },
...
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Sign in",
        "description": "One sign-in for everyone: the supervisor gets the panel session (kind admin), a user gets their own page (kind user, a separate cookie). Needs the X-Meridian: 1 header. With two-factor sign-in, the first answer is {totp_required: true}; send the same request again with code.",
        "operationId": "postLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "x-scope": "public"
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Sign out",
        "operationId": "postLogout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/meta": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "Panel name, and for signed-in callers version and protocol catalogue",
        "operationId": "getMeta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "x-scope": "public"
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "The signed-in account",
        "operationId": "getMe",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/me/password": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Change your password",
        "description": "Signs out every other session.",
        "operationId": "postMePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/me/totp/setup": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Start two-factor setup",
        "operationId": "postMeTotpSetup",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpSetup"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/me/totp/enable": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Turn on two-factor sign-in",
        "operationId": "postMeTotpEnable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpEnable"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/me/totp/disable": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Turn off two-factor sign-in",
        "operationId": "postMeTotpDisable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordOnly"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/me/sessions": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "Browsers signed in to your account",
        "operationId": "getMeSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/me/sessions/revoke-others": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Sign out every other browser",
        "operationId": "postMeSessionsRevokeOthers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/tokens": {
      "get": {
        "tags": [
          "API tokens"
        ],
        "summary": "Your API tokens",
        "operationId": "getTokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIToken"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      },
      "post": {
        "tags": [
          "API tokens"
        ],
        "summary": "Create an API token",
        "description": "The token is returned once. Send it as Authorization: Bearer \u003ctoken\u003e.",
        "operationId": "postTokens",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenCreated"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/tokens/{id}": {
      "delete": {
        "tags": [
          "API tokens"
        ],
        "summary": "Revoke an API token",
        "operationId": "deleteTokensId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "x-scope": "session"
      }
    },
    "/api/overview": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Dashboard numbers and alerts",
        "operationId": "getOverview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
...
    "/api/ips": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "IP history",
        "description": "Client IPs that connected, aggregated per IP, most recent first (at most 1000). q matches an IP prefix, organisation, city or country code.",
        "operationId": "getIps",
        "parameters": [
          {
            "description": "How many days back",
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Search text",
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only this server",
            "in": "query",
            "name": "server",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IPHistoryRow"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/dests": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Destinations",
        "description": "Where traffic went, by bytes then connections (at most 500). Bytes are exact for WireGuard; Xray and Hysteria2 report connections.",
        "operationId": "getDests",
        "parameters": [
          {
            "description": "How many days back",
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Search text",
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only this server",
            "in": "query",
            "name": "server",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Only this user",
            "in": "query",
            "name": "user",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Destination"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "summary": "Activity timeline, newest first",
        "operationId": "getEvents",
        "parameters": [
          {
            "description": "Only events with a smaller id (paging)",
            "in": "query",
            "name": "before",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Only this server",
            "in": "query",
            "name": "server",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Only this user",
            "in": "query",
            "name": "user",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "warn = warnings only",
            "in": "query",
            "name": "level",
            "schema": {
              "enum": [
                "warn"
              ],
              "type": "string"
            }
          },
          {
            "description": "1-500, default 100",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/servers": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "List servers",
        "operationId": "getServers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Server"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      },
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Add a server",
        "description": "Returns the install command to run on the server. protocols lists what to set up once the agent connects.",
        "operationId": "postServers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetail"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
...
    "/api/servers/{id}/rotate-token": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Issue a new agent token",
        "description": "The agent cannot reach the panel until it is reinstalled with the new command. Traffic is not affected.",
        "operationId": "postServersIdRotateToken",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallResult"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/servers/{id}/actions": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Restart or upgrade",
        "description": "Queues an explicit maintenance action. restart_pending restarts exactly what waits for a restart (the server's pending_restart) and disconnects those users for a moment; restart_xray and upgrade_xray disconnect Xray users for a moment; upgrade_hysteria and upgrade_realm switch the server to the version in Settings and restart those cores (their users reconnect); upgrade_agent disconnects nobody.",
        "operationId": "postServersIdActions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRef"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
    "/api/actions/{id}": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "Result of an action",
        "operationId": "getActionsId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionStatus"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "in": "cookie",
        "name": "mrd_s",
        "type": "apiKey"
      },
      "token": {
        "description": "API token from Settings › API \u0026 MCP",
        "scheme": "bearer",
        "type": "http"
      },
      "user": {
        "description": "A user's session on their own page",
        "in": "cookie",
        "name": "mrd_u",
        "type": "apiKey"
      }
    },
    "schemas": {
      "LoginInput": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "OkResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "PasswordChange": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string"
          },
          "new": {
            "description": "10 to 72 bytes",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "description": "What went wrong, in plain English",
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "token": []
    },
    {
      "session": []
    }
  ]
}